requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.2.2
django-cors-headers==3.5.0
django-heroku==0.3.1
django-mathfilters==1.0.0
django-taggit==1.3.0
gunicorn==20.1.0
idna==2.10
joblib==1.0.1
numpy==1.20.2
oauthlib==3.1.0
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.17.1
python-usernames==0.3.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
scikit-learn==0.24.1
scipy==1.6.2
sentry-sdk==1.1.0
six==1.15.0
sklearn==0.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
threadpoolctl==2.1.0
urllib3==1.26.2
validate-email==1.3
whitenoise==5.2.0
